+++ biz/error.go
@@ -1,7 +1,0 @@
-package biz
-
-import "errors"
-
-var (
-	AmountIllegal = errors.New("amount illegal")
-)