+++ config/config.go
@@ -11,7 +11,6 @@ const (
 	LogLevel              = "LOG_LEVEL"
 	LarkAppId             = "LARK_APP_ID"
 	LarkAppSecret         = "LARK_APP_SECRET"
-	LarkAppToken          = "LARK_APP_TOKEN"
 	LarkVerificationToken = "LARK_APP_VERIFICATION_TOKEN"
 )
 
@@ -23,7 +22,6 @@ type Config struct {
 type LarkConfig struct {
 	AppId             string
 	AppSecret         string
-	AppToken          string
 	VerificationToken string
 }
 
@@ -34,12 +32,10 @@ func Load() *Config {
 
 	_ = v.BindEnv(LarkAppId)
 	_ = v.BindEnv(LarkAppSecret)
-	_ = v.BindEnv(LarkAppToken)
 	_ = v.BindEnv(LarkVerificationToken)
 
 	cfg.LarkConfig.AppId = v.GetString(LarkAppId)
 	cfg.LarkConfig.AppSecret = v.GetString(LarkAppSecret)
-	cfg.LarkConfig.AppToken = v.GetString(LarkAppToken)
 	cfg.LarkConfig.VerificationToken = v.GetString(LarkVerificationToken)
 	if l, err := logrus.ParseLevel(v.GetString(LogLevel)); err == nil {
 		cfg.LogLevel = l