+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0709
+
+- 支持绑定自定义多维表格账本
+- 引入[feishu-bitable-db](https://github.com/geeklubcn/feishu-bitable-db)处理repo操作
+
 # 0703
 
 - 飞书记账机器人